cross-platform prevent goto command executed before node is launched;

+++ portablizer.pyqt4.py
@@ -1,7 +1,10 @@
 import os
 import sys
+import pkg_resources # make PyInstaller import this
 import psutil
 import subprocess
+import time
+import threading
 
 from PyQt4.QtCore import *
 from PyQt4.QtGui import *
@@ -28,7 +31,7 @@ class Portablizer(QMainWindow):
 if getattr(sys, 'frozen', False):
   basedir = sys._MEIPASS
 else:
-  basedir = os.path.dirname(__file__)
+  basedir = os.path.dirname(os.path.abspath(__file__))
 
 def nodedir(file):
   return os.path.join(basedir, "node", file)
@@ -41,12 +44,7 @@ app.setApplicationName('Portablizer')
 portablizer = Portablizer()
 portablizer.setWindowTitle('Portablizer')
 
-try:
-  startupinfo = subprocess.STARTUPINFO()
-  startupinfo.dwFlags |= subprocess.STARTF_USESHOWWINDOW
-except:
-  startupinfo = None
-
+# kill previous launched but not ended node process
 for proc in psutil.process_iter():
   try:
     if proc.name == "node.exe" and proc.getcwd() == basedir:
@@ -54,19 +52,40 @@ for proc in psutil.process_iter():
   except:
     pass
 
+# hide console window only in Windows exe
+startupinfo = None
 try:
-  nodeprocess = subprocess.Popen([nodedir("node.exe"), nodedir("app.js")], startupinfo=startupinfo, stdout=subprocess.PIPE)
+  startupinfo = subprocess.STARTUPINFO()
+  startupinfo.dwFlags |= subprocess.STARTF_USESHOWWINDOW
+except:
+  pass
+
+# can't use stdout in Windows because there is no console
+nodeprocess = None
+try:
+  subp = [nodedir("node.exe"), nodedir("app.js")]
+  if startupinfo is not None:
+    nodeprocess = subprocess.Popen(subp, startupinfo=startupinfo)
+  else:
+    nodeprocess = subprocess.Popen(subp, stdout=subprocess.PIPE)
 except:
   QMessageBox.warning(portablizer, "Error", "The application is missing required files. You need to re-install it.")
+  if hasattr(nodeprocess, 'kill'):
+    nodeprocess.kill()
   sys.exit(1)
 
 portablizer.show()
 portablizer.startBrowser()
 
-for line in iter(nodeprocess.stdout.readline, ''):
-  if "3000" in line:
-    portablizer.goto(nodeurl)
-    break
+if startupinfo is not None:
+  # in Windows, delay one second
+  threading.Timer(1, portablizer.goto(nodeurl))
+else:
+  # in unix, read node output, when server is ready, goto page
+  for line in iter(nodeprocess.stdout.readline, ''):
+    if "3000" in line:
+      portablizer.goto(nodeurl)
+      break
 
 menubar = portablizer.menuBar()
 helpMenu = menubar.addMenu('&Options')
@@ -76,9 +95,7 @@ helpMenu.addAction(openweb)
 
 appprocess = app.exec_()
 
-try:
+if hasattr(nodeprocess, 'kill'):
   nodeprocess.kill()
-except:
-  pass
 
 sys.exit(appprocess)
